+++ R/getAdjusted.R
@@ -1,7 +1,7 @@
 #' Adjust direct estimates and their associated variances
 #' 
 #' @param data data frame of the adjusted estimates and the associated uncertainties, see the arguments below for specific columns.
-#' @param ratio the ratio of unadjusted mortality rates to the true mortality rates. It can be either a data frame with three columns (region, time, and ratio) if adjustment factor differ by region; or a data frame with two columns (time and ratio) if adjustment factor only varies over time.
+#' @param ratio the ratio of unadjusted mortality rates to the true mortality rates. It can be either a data frame with the following three columns (region, time, and ratio) if adjustment factor differ by region; or a data frame with the following two columns (time and ratio) if adjustment factor only varies over time.
 #' @param time the column name for time in the data 
 #' @param region the column name for region in the data 
 #' @param est the column name for unadjusted mortality rates in the data 
@@ -13,8 +13,27 @@
 #' 
 #' @return adjusted dataset of the same columns.
 #' @examples
+#' \dontrun{
+#' years <- levels(DemoData[[1]]$time)
 #' 
-#'  
+#' # obtain direct estimates
+#' data <- getDirectList(births = DemoData, 
+#' years = years,  
+#' regionVar = "region", timeVar = "time", 
+#' clusterVar = "~clustid+id", 
+#' ageVar = "age", weightsVar = "weights", 
+#' geo.recode = NULL)
+#' # obtain direct estimates
+#' data_multi <- getDirectList(births = DemoData, years = years,
+#'   regionVar = "region",  timeVar = "time", clusterVar = "~clustid+id",
+#'   ageVar = "age", weightsVar = "weights", geo.recode = NULL)
+#' data <- aggregateSurvey(data_multi)
+#' 
+#' # randomly simulate adjustment factor
+#' adj <- expand.grid(region = unique(data$region), time = years)
+#' adj$ratio <- runif(dim(adj)[1], min = 0.5, max = 0.8)
+#' data.adj <- getAdjusted(data = data, ratio = adj)
+#'  }
 #' @export
 #'